requirements.txt
discord.py==1.7.1
python-dotenv==1.0.1